+++ wdtk-datamodel/src/main/java/org/wikidata/wdtk/datamodel/json/jackson/datavalues/JacksonInnerGlobeCoordinates.java
@@ -1,8 +1,11 @@
 package org.wikidata.wdtk.datamodel.json.jackson.datavalues;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.wikidata.wdtk.datamodel.interfaces.GlobeCoordinatesValue;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
 
 /*
  * #%L
@@ -35,9 +38,12 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class JacksonInnerGlobeCoordinates {
 
+	static final Logger logger = LoggerFactory
+			.getLogger(JacksonInnerGlobeCoordinates.class);
+
 	private double latitude = 0;
 	private double longitude = 0;
-	private double precision = 0;
+	private double precision = GlobeCoordinatesValue.PREC_ARCSECOND;
 	private String globe = GlobeCoordinatesValue.GLOBE_EARTH;
 
 	/**
@@ -45,22 +51,6 @@ public class JacksonInnerGlobeCoordinates {
 	 * deserialization. Should only be used by Jackson for this very purpose.
 	 */
 	public JacksonInnerGlobeCoordinates() {
-	}
-
-	/**
-	 * TODO Review the utility of this constructor.
-	 *
-	 * @param latitude
-	 * @param longitude
-	 * @param precision
-	 * @param globe
-	 */
-	public JacksonInnerGlobeCoordinates(double latitude, double longitude,
-			double precision, String globe) {
-		this.latitude = latitude;
-		this.longitude = longitude;
-		this.precision = precision;
-		this.globe = globe;
 	}
 
 	/**
@@ -111,8 +101,9 @@ public class JacksonInnerGlobeCoordinates {
 	 * @see GlobeCoordinatesValue#getPrecision()
 	 * @return precision
 	 */
-	public double getPrecision() {
-		return this.precision;
+	@JsonProperty("precision")
+	public double getJsonPrecision() {
+		return this.precision / GlobeCoordinatesValue.PREC_DEGREE;
 	}
 
 	/**
@@ -122,8 +113,17 @@ public class JacksonInnerGlobeCoordinates {
 	 * @param precision
 	 *            new value
 	 */
-	public void setPrecision(double precision) {
-		this.precision = precision;
+	@JsonProperty("precision")
+	public void setJsonPrecision(double precision) {
+		if (precision <= 0.0) {
+			// We just do this silently because it is so common in the data.
+			// Precision "0" does not make sense for a physical quantity.
+			// Automatic precision does not make sense for floating point
+			// values. "0" also is commonly produced from "null" in JSON.
+			this.precision = GlobeCoordinatesValue.PREC_ARCSECOND;
+		} else {
+			this.precision = precision;
+		}
 	}
 
 	/**